Basic cue merge framework sketched out.
Now, unlike with the ramp-channels function, the plan is to have a
single scheduled task which looks at a map of active cues/effects, whose
keys can be used to update or remove them, and whose values are
functions called for each DMX frame, which return a list of universe,
channel and DMX value tuples to take effect on that frame. As a basic
proof of concept, a cue which sets the RGB values of any patched
fixtures passed in to it can be added and removed, with visible results.
Onward now to time-varying functions and composite fun!