Use more basic thread handoff from MIDI events.
Apparently core.async uses too many tricks to be efficient with threads,
which was defeating my purpose in trying to take MIDI event processing
away from the MIDI extension threads which have classloader issues. Use
a basic BlockingQueue and my own dedicated thread to be darn sure. This
change makes it so I can start a show for the first time using the Stop
button on the Ableton Push, and everything finally works. Whew!