Fix many issues with dynamic spatial params.
I had previously only been testing static spatial params (which were
precomputed when created, not changing during the course of time), and
there were many problems with the implementation of the latter. I also
changed the design and documentation in order to better support the
kinds of things it turns out you want to do when creating moving head
chases.