Add held cue feedback, shift key override.
Now, cues are only whitened while held down if they are going to
self-terminate when released (i.e. have the :held flag set). Also, the
show operator can override the :held flag by holding down the shift key
on either the web interface or the Ableton Push when activating the cue.
Doing so causes it to act as a normal cue, which runs until pressed
again.