Upgrade cue aftertouch support to general velocity
Cues which respond to aftertouch pressure need to also respond to the
velocity of the initial Note On event which launches them, or there is a
glaring visual discontinuity. This also simpifies the concept, and
allows the keywords and description to refer to the more-familiar
concept of velocity rather than aftertouch, which will make it
accessible to more users. Also, controllers which support only velocity
but not aftertouch can now at least offer that feature.