Add shutdown hooks for graceful cleanup.
Now when the Java environment is shutting down, any running shows are
stopped and blacked out, and any active Ableton Push controllers are
deactivated.

This change also switched from using at-at to schedule each DMX frame
update to using a dedicated thread in an infinite loop, which is ready
to implement a callback function for afterglow-max to be notified in
advance of the metronome snapshot of the next frame that will be sent,
so Max patchers can set variables appropriately.