Changes to support non-local OLA (for Windows).
Add ability to specify a different OLA server port and host. Since this
could be slower and more error prone, add socket timeout (and connection
timeout) limits, of 2 seconds. Also beef up communication loops to be
more robust in cases of communication failure.

Also added a note that the MIDI message dispatch thread is a bit too
trusting that its handlers will not block it, so it should be revisited.