Flesh out command-line options.
Now you can specify the address and port on which to find the OLA
daemon, the path to the log file, and a list of initialization files to
be loaded at startup as command-line arguments, as well as explicitly
ask for help. This will make it a lot easier to run Afterglow as an
executable jar under different circumstances, and set up your own custom
shows.