Simplify tracking of MIDI clock sources.
Tracking is always performed when MIDI events are being received, so
there is no longer a need to pause before displaying user interfaces
which want to offer sync choices. And the detection of Traktor beat
phase information is now more reliable, so it no longer needs to pay
attention to the name of the MIDI port it is coming in on.