Add integration with Pangolin Beyond laser shows
If you turn on Beyond's UDP PangoScript Talk server, you can tell
Afterglow the address and port on which to find it, and sync the BPM and
beat grid of the laser show to the main metronome of an Afterglow show.
You can also send other commands to trigger cues and adjust effects; I
will explore some of that next.