Setting up for more formal rendering extensions.
Establish a general mechanism for new kinds of effects and assigners,
which do not result in DMX values being written to the show's universes,
to nonetheless participate in Afterglow's rendering loop. This will
replace the hacky approach used last weekend to allow integration with
Pangolin Beyond laser shows, since that would not have worked properly
with fades, for example. Once this is finished, remove the changes that
were made to try to let effects know when they were ending; that is one
thing that would not work with fades, so it should not even pretend to
be supported.