Add extension framework to rendering loop.
Rewrite the formerly hacked together Beyond laser show integration to
work with this new extension framework, so laser effects can be actual
first class participants in the show rendering process. Now they will be
able to participate in fades and chases as well, and I can go back and
remove the ill-conceived attempt to inform effects when they are going
to end.

Also further increased parallelism of the rendering loop, and split out
some functions for improved readability. Fixed a race condition in
checking whether effects should end (which would cause problems when
effects were ended by the user at the same time).

Updated the profiling function to work with the new extension mechanism,
and made it able to run both with and without a thread pool, in order to
allow analysis of the effectiveness of the concurrency.