Remove ill-conceived effect end feedback attempt
This commit rolls back the changes which tried to ensure that an
effect's end function was called at least once when it was going to no
longer be used. They were added as part of the first approach to
supporting new kinds of effects (specifically the Pangolin Beyond laser
show integration) which do not generate values for the show's DMX
universes. However, it turns out there is no reliable way to always let
an effect know it is ending, particularly when it is involved in chases
and fades between effects.

These changes became unnecessary when the more general mechanism for
extending the rendering loop with new effect types was designed and
implemented. There were a couple of variable renamings which are
improvements and are being retained.

The documentation was updated to make it clear that effects must not
rely on their end function being called.