Add an effect which transforms color effects.
This is a generalization of the "rainbow pulse" effect which started out
in afterglow-max and was ported to the example show. Rather than being a
single effect, this creates a family of effects which can modify the
colors being sent to the fixtures they are assigned. By default, it will
cause the colors to change saturation over the course of each beat, from
full saturation at the start to no saturation by the end. But it is easy
to create other transformation functions using oscillated parameters.