Have chase delegate still-active?/end
The previous implementation discarded any notion of tracking the ending
of underlying effects in the chase; they were assumed to run forever
until the chase itself was ended. Now, they are asked if they are
running, and replaced with blank effects if they end themselves before
the chase itself ends. When the chase is asked to end, it now asks each
constituent effect to end, and will continue running until they all have
done so.