Restore useful stack traces for web runtime errors
It turns out the Prone library which formats them was unable to parse
the project file once that started including unescaped forms to
evaluate, and it would crash trying to format the error that it was
supposed to display. By telling it the application namespace prefix, we
can avoid it need to try to read the project file at all.