Environment changes working, #38
Have both addition and removal of devices working smoothly with the
unreleased version of CoreMIDI4J! A few loose ends and improvements
remain to be made in midi.clj, but this is already fantastic!

Also still want to add examples of watching for and auto-mapping devices
to automate recovery from interrupted connections.