Work better with standard MIDI SPI.
It turns out that the standard MIDI implementation that ships with Java
returns brand new MidiDevice objects every time you ask for a list of
them, which completely wrecked Afterglow's ability to keep track of
device changes. However, it seems that the devices' underlying
MidiDevice.Info objects do *not* change, and that is also true with
CoreMidi4J, so changing to use them as keys works much better.