Give visual feedback about *all* cue conflicts!
Woot, it turned out to be really easy to have grid controllers dim out
all cues which would either be killed by or kill a cue that is
running (in other words, taking into account the :end-keys list as well
as the cue :key). Now the grid gives you an even more intuitive sense of
the relatedness of cues.