Move API docs to rawgit.
By building the codox documentation to an api-doc subdirectory within
the branch itself, rather than copying it to a separate Github Pages
directory for each release, it can be kept right inside the tag that
corresponds with a release, and it can be delivered via the rawgit
system and properly rendered as web pages. This is a simpler, safer
approach that what I thought I would need to use.