Better handle velocity launching cue grid cues.
Previously it was up to the controller to try to interpret the velocity
and pass in the resulting cue variables as overrides. That doesn't make
sense since it duplicates code, and resulted in unbound cue variables
when launching from the web interface. Now we explicitly take a velocity
value and handle it when launching the cue, defaulting to 127 if not
specified.