Consolidate cleanup of active controller bindings.
Each of the MIDI grid controller bindings was maintaining its own list
to be shut down when the JVM was ending, or when deactivate-all was
called, and was doing so with an unnecessary id counter. Now the
controllers namespace manages a set of functions to be called to shut
down active bindings on behalf of all controller mappings.