Wrap all MIDI event handlers uniformly.
Not all registered MIDI event handlers were being called inside of a
try/catch block or being checked for thread interruption. This change
uses a single delegation function to do that in a uniform way, and also
offers a single point where they can be moved onto another thread to
protect against resource starvation if and when I am ready to do that.

This change also fixes the API doc describing the mapping atoms, it was
out of date.