Simplify working with color cue vars.
The function which helps you display the current color in the UI now
understands saved cue variables. Simplified the example color cues in
to take advantage of this, and get rid of the permanent show variables
they had been using to save color values before this week's cue saving
mechanism was implemented.