Improve controller tempo tap and documentation.
The low-level tempo tap handler was already more useful than I was
giving it credit for, suitable for both aligning to the current beat as
well as adjusting the tempo if you hit it three or more times, so the
shift key can be used to adjust the down beat even on unsynchronized
shows. This makes it much easier to keep the lights in sync manually!