Replace cue var slider tooltips with labels.
The tooltips got in the way of dragging other cue variables because they
pop up even when you are not on the slider track. If/when that issue in
the upstream project is fixed, we can consider switching back to using
them, but even then they are not particularly helpful on an iPad because
there is no way to hover over the track to bring them up.