Remove spurious MIDI updates.
Comparing the previous and current text display lines was not being done
right, so the entire Push display was being sent on each frame even when
nothing had changed. Similar issues were causing all of the text buttons
have their state set on each frame. Now comparison is done properly, so
only when a display line or button state has changed is a MIDI message
sent to update it on the Push.