Flesh out metronome adjustment UI.
Add a monospace font so the numbers stay stable as their values cycle.
Fix the way that string widths are being calculated to work with custom
fonts. Add highlights to indicate the portion of the BPM or beat value
which is beign adjusted. Pass the display frame snapshot into the
interface overlays so they can all work with the same notion of the
current time.