Support prioritization of overlays.
This will allow the Show Stopped overlay on the Push 2 to float over
even more recent overlays if the user adjusts cue variables while the
show is stopped. Also found a more idiomatic and efficient way of
keeping the map sorted in the right order, so the reverse step can be
eliminated when running the overlays.