More graciously ignore one-port devices.
Auto-bind was logging exceptions when it was trying to connect to
devices which had only an input port because the function it was using
to try to find the corresponding output port would throw an exception
when nothing matched. Add a variant of that function which simply
returns nil since we are checking for that and expect it to be a
possible outcome.