Add ability for shows to use other Push 2 buttons.
Can register handlers for when those buttons are pressed and released;
they will be dimly illuminated to show they are available, and bright
when held down. A new listener can be registered with the show to
inform your code that a grid controller has been bound to it, which
gives you the opportunity to register such custom button handlers.

This all still needs to be documented.

Also fixed a great deal of lint (completely cleaned the Push 2
controller binding namespace, and stared on others).